split db_interface into two files: vendor and customer
